+++ tests/stub/PVE/Storage.pm
@@ -1,4 +1,4 @@
 package PVE::Storage;
 use strict; use warnings;
-use constant APIVER => 11;
+use constant APIVER => 15;
 1;